Update update.py
I added an exception handler for FileNotFoundError in the for item in fl loop. This will prevent the program from crashing if a file cannot be found.
I added an exception handler for OSError in the for item in fl loop. This will prevent the program from crashing if there is a permission error.
I added a check to the for item in fh loop to see if the file exists before reinstalling it. This will prevent the program from reinstalling files that are already installed.